server/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
chromadb
PyPDF2
python-docx
requests
tqdm
tiktoken
python-dotenv
pytest
typing_extensions
black
flake8
isort
pytesseract
Pillow
openai-whisper
moviepy